CVE-2021-47257: Add .vulnerable file
The root cause of the bug was introduced in commit
a26c5fd7622d4951425131d54a8c99f076fe2068 by Alexander Aring on September
28, 2015, when the ieee802154_llsec_parse_dev_addr function was first
added to the kernel.

The function was originally introduced with a flawed validation logic
that used
AND (&&) operators, which meant the validation would only fail if ALL
three conditions were false. This was clearly wrong - it should fail if
ANY required attribute is missing.

Even after a partial fix in commit 421eeea10d0a5 that changed AND to
OR, the validation logic was still incorrect. It required that EITHER a
SHORT or EXTENDED address attribute must be present, regardless of the
mode. But then it would access the specific attribute based on the mode
without checking if that specific attribute exists.

The bug created a null pointer dereference vulnerability. An attacker
could send a netlink message with:
- NL802154_DEV_ADDR_ATTR_PAN_ID set
- NL802154_DEV_ADDR_ATTR_MODE set to NL802154_DEV_ADDR_SHORT
- NL802154_DEV_ADDR_ATTR_EXTENDED present (but not
  NL802154_DEV_ADDR_ATTR_SHORT)

The validation would pass, but when executing the switch statement
based on mode
NL802154_DEV_ADDR_SHORT, it would try to access
attrs[NL802154_DEV_ADDR_ATTR_SHORT] which is NULL, causing
nla_get_le16(NULL) to trigger a null pointer dereference.

The bug persisted through a partial fix and wasn't fully resolved until
commit
9fdd04918a452980631ecc499317881c1d120b70, approximately 5.5 years after
its introduction. The vulnerability could allow local attackers to crash
the kernel through a crafted netlink message that passes validation but
causes a null pointer dereference when processing the address based on
the specified mode.

Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>