tools: voting_results: add --no-annotate option
When assigning CVEs, I really don't want the output to show all the
annotations as it gets in the way of the output that cve_batch_create
wants.  So add an option, that is not set by default, to not show the
annotations so they don't get in the way.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>