tools: dyad: handle issue when fixes go back in time but vulnerability did not
Sometimes a fix is backported to kernels where the vulnerable issue was
not backported.  Previously we would record this "range" as a vulnerable
range, despite it not being a correct range at all, just to be
"correct".  This causes problems and is obviously confusing.

Fix this all up by not recording ranges like this at all, as obviously
there was not a vulnerable range so saying there was is not ok.

Update the test that was previously checking for this bizarre situation,
so that if we change this in the future it also will need to be changed
(i.e. so we don't break anything.)

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>