Set a lower bound for CVE-2023-53431
In 54a99fcd011c69b ("fix the git id for CVE-2023-53431") the fix id was set to
a commit that reverted the original fix attempt and used a different approach.
However that commit obviously contains a Fixes: tag: the original fix attempt.
So now suddenly the issue was flagged as existing only between the first and
second fixes.

Before that it had no start version set, i.e. "forever". Use the commit that
originally introduced the enclosure ULD as lower barrier instead, which is
likely more correct than any of the previous start versions. This commit is not
necessarily where the issues were actually introduced, it just serves as a
lower barrier to denote that no older versions can be affected, and that the
Fixes: tag from the second fix isn't the true story either.

Signed-off-by: Rolf Eike Beer <eb@emlix.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>