dyad: reduce cloning in filter_and_sort_pairs
Change filter_and_sort_pairs to take ownership of the input Vec
instead of borrowing it. This allows pairs to be moved rather
than cloned when grouping by vulnerability ID and when adding
to the filtered output.

Also consolidate the two-pass mainline detection into a single
iteration that processes pairs as they're consumed.

Signed-off-by: Sasha Levin <sashal@kernel.org>