CVE-2023-52811: Add .vulnerable file
The vulnerability (CVE-2023-52811) was introduced in commit
072b91f9c651 ("[SCSI] ibmvfc: IBM Power Virtual Fibre Channel Adapter
Client Driver").

Commit details:
- Commit Hash: 072b91f9c6510d0ec4a49d07dbc318760c7da7b3
- Author: Brian King <brking@linux.vnet.ibm.com>
- Date: July 1, 2008
- Kernel Version: v2.6.27-rc1
- Affected File: drivers/scsi/ibmvscsi/ibmvfc.c
- Affected Function: ibmvfc_get_event()

The vulnerability was introduced in the original ibmvfc driver
implementation in 2008. The developer used
`BUG_ON(list_empty(&vhost->free))` as an assertion that the event pool
should never become empty. This was a programming pattern that was
acceptable at the time but is now recognized as dangerous because:

1. It causes a kernel crash for a potentially recoverable condition
2. If the kernel continues past the panic, it would access invalid
   memory from an empty list
3. It provides no graceful degradation or error recovery path

Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>